+++ indexing-service/src/main/java/org/apache/druid/indexing/common/task/InputSourceProcessor.java
@@ -35,7 +35,6 @@ import org.apache.druid.java.util.common.logger.Logger;
 import org.apache.druid.java.util.common.parsers.CloseableIterator;
 import org.apache.druid.java.util.common.parsers.ParseException;
 import org.apache.druid.query.aggregation.AggregatorFactory;
-import org.apache.druid.segment.SegmentUtils;
 import org.apache.druid.segment.indexing.DataSchema;
 import org.apache.druid.segment.indexing.granularity.GranularitySpec;
 import org.apache.druid.segment.realtime.appenderator.AppenderatorDriverAddResult;
@@ -156,7 +155,7 @@ public class InputSourceProcessor
                 // If those segments are not pushed here, the remaining available space in appenderator will be kept
                 // small which could lead to smaller segments.
                 final SegmentsAndCommitMetadata pushed = driver.pushAllAndClear(pushTimeout);
-                SegmentUtils.logSegments(LOG::debug, pushed.getSegments(), "Pushed segments");
+                LOG.debugSegments(pushed.getSegments(), "Pushed segments");
               }
             }
           } else {
@@ -175,7 +174,7 @@ public class InputSourceProcessor
       }
 
       final SegmentsAndCommitMetadata pushed = driver.pushAllAndClear(pushTimeout);
-      SegmentUtils.logSegments(LOG::debug, pushed.getSegments(), "Pushed segments");
+      LOG.debugSegments(pushed.getSegments(), "Pushed segments");
 
       return pushed;
     }
