+++ server/src/main/java/org/apache/druid/server/coordinator/duty/CompactSegments.java
@@ -29,7 +29,6 @@ import org.apache.druid.client.indexing.TaskPayloadResponse;
 import org.apache.druid.indexer.TaskStatusPlus;
 import org.apache.druid.java.util.common.ISE;
 import org.apache.druid.java.util.common.logger.Logger;
-import org.apache.druid.segment.SegmentUtils;
 import org.apache.druid.server.coordinator.CoordinatorCompactionConfig;
 import org.apache.druid.server.coordinator.CoordinatorStats;
 import org.apache.druid.server.coordinator.DataSourceCompactionConfig;
@@ -212,7 +211,7 @@ public class CompactSegments implements CoordinatorDuty
             taskId,
             segmentsToCompact.size()
         );
-        SegmentUtils.logSegments(LOG::info, segmentsToCompact, "Compacting segments");
+        LOG.infoSegments(segmentsToCompact, "Compacting segments");
         // Count the compaction task itself + its sub tasks
         numSubmittedTasks += findNumMaxConcurrentSubTasks(config.getTuningConfig()) + 1;
       } else {
